Add new Wall Behaviour

+++ inc/Entity.hpp
@@ -6,7 +6,7 @@
 /*   By: tchartie <tchartie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/13 23:00:21 by tchartie          #+#    #+#             */
-/*   Updated: 2025/05/27 17:58:28 by tchartie         ###   ########.fr       */
+/*   Updated: 2025/08/01 16:25:00 by tchartie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,18 +20,18 @@ class Entity {
 		Entity( void );
 		virtual ~Entity( void );
 
-		size_t	getPosX( void );
-		size_t	getPosY( void );
+		int		getPosX( void );
+		int		getPosY( void );
 		size_t	getLife( void );
 		size_t	getMaxLife( void );
 
-		void	setPosX( size_t newX );
-		void	setPosY( size_t newY );
+		void	setPosX( int newX );
+		void	setPosY( int newY );
 		void	setLife( size_t newLife );
 		void	setMaxLife( size_t newMaxLife );
 	protected:
-		size_t	_posX;
-		size_t	_posY;
+		int		_posX;
+		int		_posY;
 		size_t	_life;
 		size_t	_maxLife;
 };
